프린터 예약 시 연속되게 예약하는 게 불가한 이슈 & 자신만 예약 제거 가능하도록 하는 이슈 해결  (#35)
* fix: 프린터 예약 시 연속되게 예약하는 게 불가한 이슈 해결

* improvement: 자신만 프린터 예약 지울 수 있도록 함

+++ src/printer-reservation/printer-reservation.service.ts
@@ -1,4 +1,8 @@
-import { BadRequestException, Injectable } from '@nestjs/common'
+import {
+    BadRequestException,
+    ForbiddenException,
+    Injectable,
+} from '@nestjs/common'
 import { User } from '@prisma/client'
 import { addDays, addHours } from 'date-fns'
 import { PrismaService } from 'nestjs-prisma'
@@ -33,12 +37,12 @@ export class PrinterReservationService {
                         {
                             requestStartTime: {
                                 gte: startTime,
-                                lte: endTime,
+                                lt: endTime,
                             },
                         },
                         {
                             requestEndTime: {
-                                gte: startTime,
+                                gt: startTime,
                                 lte: endTime,
                             },
                         },
@@ -113,9 +117,23 @@ export class PrinterReservationService {
         })
     }
 
+    async checkOwnershipOfReservation(
+        reservationId: number,
+        userId: User['id'],
+    ) {
+        const reservation = await this.prisma.printerReservation.findUnique({
+            where: { id: reservationId },
+        })
+        if (reservation?.userId !== userId) {
+            throw new ForbiddenException('Unauthorized User')
+        }
+    }
+
     async deleteReservationById(
         reservationId: number,
+        userId: User['id'],
     ): Promise<PrinterReservationDto> {
+        await this.checkOwnershipOfReservation(reservationId, userId)
         const reservationToDelete = await this.prisma.printerReservation.delete(
             {
                 where: {
